requirements.txt
os
datetime
pathlib
math
functools
re
numpy==1.23.5
scipy==1.10.1
pandas==1.5.3
tifffile==2023.4.12
opencv-python==4.9.0.80
matplotlib==3.7.1
scikit-image==0.19.3
antspyx==0.3.8
tqdm==4.65.0
